Project-3/project/requirements.txt
streamlit==1.46.0
tensorflow
pillow
numpy
